+++ chat-widget/src/components/webchatcontainerstateful/common/defaultProps/defaultMiddlewareLocalizedTexts.ts
@@ -32,5 +32,6 @@ export const defaultMiddlewareLocalizedTexts: ILiveChatWidgetLocalizedTexts = {
     MIDDLEWARE_BANNER_FILE_EMAIL_ADDRESS_RECORDED_SUCCESS: "Email will be sent after chat ends!",
     MIDDLEWARE_BANNER_FILE_EMAIL_ADDRESS_RECORDED_ERROR: "Email {0} could not be saved, try again later.",
     PREVIOUS_MESSAGES_LOADING: "Loading previous messages...",
-    CONVERSATION_DIVIDER_ARIA_LABEL: "Conversation history divider"
+    CONVERSATION_DIVIDER_ARIA_LABEL: "Conversation history divider",
+    INITIAL_HISTORY_LOADING: "Loading previous messages..."
 };